requirements.txt
requests==2.31.0
pyodbc==4.0.31
python-dotenv==0.21.1
pytest==7.2.1
pytest-cov==4.0.0
pre-commit==2.21.0
pylint==2.17.1
pandas==1.5.3
sqlalchemy==1.4.46
googleads==38.0.0
google-api-core==2.8.0
google-auth==2.23.4
google-auth-oauthlib==0.8.0
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.3
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
libcst==0.3.20
pyarrow==14.0.1
detect_secrets==1.4.0
pandas-gbq==0.13.2
google-cloud-storage==2.1.0
retrying==1.3.4
fsspec==2023.9.2
black==23.10.1
pydantic==2.5.2
protobuf==3.20.2
beautifulsoup4==4.12.2
db-dtypes
backoff
s3fs
boto3
pyyaml
ruff==0.7.2


# aiobotocore                   2.5.2
# s3fs                          2023.6.0
# boto3                         1.28.11
# fsspec                        2023.6.0
# botocore                      1.31.85


# aiobotocore                   2.5.4
# boto3                         1.28.11
# botocore                      1.31.17
# fsspec                        2023.9.2
# s3fs                          2023.9.2
aiobotocore>=2.9.1 # not directly required, pinned by Snyk to avoid a vulnerability
aiohttp>=3.9.4 # not directly required, pinned by Snyk to avoid a vulnerability
numpy>=1.22.2,<2.0.0 # not directly required, pinned by Snyk to avoid a vulnerability # Needs to stay below 2.0.0 see https://stackoverflow.com/questions/78634235/numpy-dtype-size-changed-may-indicate-binary-incompatibility-expected-96-from
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
idna>=3.7 # not directly required, pinned by Snyk to avoid a vulnerability
